move get_model_data_and_priors_from_dir to utils
