+++ pipelines/utils.py
@@ -4,8 +4,10 @@ pipeline.
 """
 
 import datetime
+import json
 import os
 import re
+import runpy
 from pathlib import Path
 
 from forecasttools import ensure_listlike, location_table
@@ -133,3 +135,17 @@ def get_all_model_run_dirs(parent_dir: Path) -> list[str]:
         for f in os.scandir(parent_dir)
         if f.is_dir() and f.name in loc_abbrs_
     ]
+
+
+def get_model_data_and_priors_from_dir(model_dir):
+    data_path = Path(model_dir) / "data" / "data_for_model_fit.json"
+    prior_path = Path(model_dir) / "priors.py"
+    priors = runpy.run_path(str(prior_path))
+
+    with open(
+        data_path,
+        "r",
+    ) as file:
+        model_data = json.load(file)
+
+    return model_data, priors
